Annotate memory leak

+++ radio/src/main/java/kz/zhombie/radio/RadioExt.kt
@@ -6,9 +6,9 @@ fun Radio.formatToDigitalClock(milliseconds: Long?): String {
     if (isReleased()) return "00:00"
     if (milliseconds == null) return "00:00"
     return try {
-        val hours = TimeUnit.MILLISECONDS.toHours(milliseconds).toInt() % 24
-        val minutes = TimeUnit.MILLISECONDS.toMinutes(milliseconds).toInt() % 60
-        val seconds = TimeUnit.MILLISECONDS.toSeconds(milliseconds).toInt() % 60
+        val hours = TimeUnit.MILLISECONDS.toHours(milliseconds) % 24
+        val minutes = TimeUnit.MILLISECONDS.toMinutes(milliseconds) % 60
+        val seconds = TimeUnit.MILLISECONDS.toSeconds(milliseconds) % 60
         when {
             hours > 0 -> String.format("%d:%02d:%02d", hours, minutes, seconds)
             minutes > 0 -> String.format("%02d:%02d", minutes, seconds)
